# src/main/resources/specifications/nws-openapi.json
  {
    "openapi": "3.0.3",
    "info": {
      "title": "weather.gov API",
      "description": "weather.gov API",
      "version": "2.6.0"
    },
    "servers": [
      {
        "url": "https://api.weather.gov",
        "description": "Production server"
      }
    ],
    "paths": {
...
      "/products/{productId}": {
        "get": {
          "description": "Returns a specific text product",
          "operationId": "product",
          "responses": {
            "200": {
              "description": "success",
              "headers": {
                "X-Correlation-Id": {
                  "$ref": "#/components/headers/CorrelationId"
                },
                "X-Request-Id": {
                  "$ref": "#/components/headers/RequestId"
                },
                "X-Server-Id": {
                  "$ref": "#/components/headers/ServerId"
                }
              },
              "content": {
                "application/ld+json": {
                  "schema": {
                    "$ref": "#/components/schemas/TextProduct"
                  }
                }
              }
            },
            "default": {
              "$ref": "#/components/responses/Error"
            }
          },
          "parameters": []
        },
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": ".",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
...
    },
    "components": {
      "schemas": {
...
        "JsonLdContext": {
          "anyOf": [
            {
              "type": "array",
              "items": {}
            },
            {
              "type": "object"
            }
          ]
        },
...
        "ProblemDetail": {
          "required": [
            "type",
            "title",
            "status",
            "detail",
            "instance",
            "correlationId"
          ],
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "description": "A URI reference (RFC 3986) that identifies the problem type. This is only an identifier and is not necessarily a resolvable URL.\n",
              "format": "uri",
              "default": "about:blank",
              "example": "urn:noaa:nws:api:UnexpectedProblem"
            },
            "title": {
              "type": "string",
              "description": "A short, human-readable summary of the problem type.",
              "example": "Unexpected Problem"
            },
            "status": {
              "maximum": 999,
              "minimum": 100,
              "type": "number",
              "description": "The HTTP status code (RFC 7231, Section 6) generated by the origin server for this occurrence of the problem.\n",
              "example": 500
            },
            "detail": {
              "type": "string",
              "description": "A human-readable explanation specific to this occurrence of the problem.",
              "example": "An unexpected problem has occurred."
            },
            "instance": {
              "type": "string",
              "description": "A URI reference (RFC 3986) that identifies the specific occurrence of the problem. This is only an identifier and is not necessarily a resolvable URL.\n",
              "format": "uri",
              "example": "urn:noaa:nws:api:request:493c3a1d-f87e-407f-ae2c-24483f5aab63"
            },
            "correlationId": {
              "type": "string",
              "description": "A unique identifier for the request, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
              "example": "493c3a1d-f87e-407f-ae2c-24483f5aab63"
            }
          },
          "description": "Detail about an error. This document conforms to RFC 7807 (Problem Details for HTTP APIs).",
          "additionalProperties": true
        },
...
        "TextProduct": {
          "type": "object",
          "properties": {
            "@context": {
              "$ref": "#/components/schemas/JsonLdContext"
            },
            "@id": {
              "type": "string",
              "format": "uri"
            },
            "id": {
              "type": "string"
            },
            "wmoCollectiveId": {
              "type": "string"
            },
            "issuingOffice": {
              "type": "string"
            },
            "issuanceTime": {
              "type": "string",
              "format": "date-time"
            },
            "productCode": {
              "type": "string"
            },
            "productName": {
              "type": "string"
            },
            "productText": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
...
      },
      "responses": {
...
        "Error": {
          "description": "An error response.",
          "headers": {
            "X-Correlation-Id": {
              "$ref": "#/components/headers/CorrelationId"
            },
            "X-Request-Id": {
              "$ref": "#/components/headers/RequestId"
            },
            "X-Server-Id": {
              "$ref": "#/components/headers/ServerId"
            }
          },
          "content": {
            "application/problem+json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemDetail"
              }
            }
          }
        },
...
      },
...
      "headers": {
        "CorrelationId": {
          "description": "A unique identifier for the request, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
          "schema": {
            "type": "string"
          }
        },
        "RequestId": {
          "description": "A unique identifier for the request, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
          "schema": {
            "type": "string"
          }
        },
        "ServerId": {
          "description": "The identifier of the server that generated the response, used for NWS debugging purposes. Please include this identifier with any correspondence to help us investigate your issue.\n",
          "schema": {
            "type": "string"
          }
        }
      },
      "securitySchemes": {
        "userAgent": {
          "type": "apiKey",
          "description": "We require that all consumers of the API include a User-Agent header in requests. This is due to a high number of scripts exhibiting abusive behavior (intentional or unintentional). We recommend setting the value to something that identifies your application and includes a contact email. This will help us contact you if we notice unusual behavior and also aid in troubleshooting issues.\nThe API remains open and free to use and there are no limits imposed based on the User-Agent string.\nThis mechanism will be replaced with a more typical API key system at a later date.\n",
          "name": "User-Agent",
          "in": "header"
        }
      }
    },
    "security": [
      {
        "userAgent": []
      }
    ],
...
  }